refactor(home): extract HomeSelect into home/home-select
